Move `CountriesSelectPresenter` into `countries` folder
We will soon have more country-related code, so move this into its own folder, analogous to our use of the `nations` folder
